+++ step_definitions/management_classes/createClass_steps.js
@@ -24,5 +24,9 @@ When(/^.* opens classes tab with name '([^']*)'$/, async function (classesTabNam
 });
 
 Then(/^.* class with name '([^']*)' and '([^']*)' members is visible$/, async function (className, membersCount) {
-	await manageClass.isNewEmptyClassCreated(className, membersCount);
+	await manageClass.isNewClassCreated(className, membersCount);
+});
+
+Then(/^.* number of students in class with name '([^']*)' is '([^']*)'$/, async function (className, membersCount) {
+	await manageClassesPage.isNumberOfMembersInClass(className, membersCount);
 });