+++ step_definitions/management_students/createStudent_steps.js
@@ -5,7 +5,6 @@ const navigationLeftPanel = require('../../page-objects/pages/NavigationLeftPage
 const elementHelpers = require('../../runtime/helpers/elementHelpers.js');
 const mailCatcher = require('../../runtime/helpers/mailCatcher.js');
 const registrationPage = require('../../page-objects/pages/generalPagesBeforeLogin/RegistrationPage');
-const { Then } = require('cucumber');
 let pin;
 //WHEN
 When(/^.* goes to students management$/, async function () {
@@ -75,22 +74,30 @@ Then(
 );
 
 Then(/^parents accept all$/, async function () {
-	return registrationPage.acceptConsent();
+	return registrationPage.acceptConsent(true);
 });
 
-Then(/^parents click on send pin code$/, async function () {
+Then(/^student accepts all$/, async function () {
+	return registrationPage.acceptConsent(false);
+});
+
+Then(/^.* click on send pin code$/, async function () {
 	return registrationPage.clickRequestPin();
 });
 
-Then(/^parent email receives 4 digit pin code$/, async function () {
+Then(/^.* email receives 4 digit pin code$/, async function () {
 	pin = await mailCatcher.getEmailPin();
 	return pin;
 });
 
-Then(/^parent fills in pin and submits$/, async function () {
+Then(/^.* fills in pin and submits$/, async function () {
 	return registrationPage.addPin(pin);
 });
 
 Then(/^login data is received$/, async function () {
 	return registrationPage.getTitleText();
+});
+
+Then(/^student selects over 16 checkbox$/, async function () {
+	return registrationPage.clickOver16Btn();
 });